% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_BASFOR_general.R
\name{run_mod_model}
\alias{run_mod_model}
\title{Run the forest model BASFOR}
\usage{
run_mod_model(rs = RESTART, ft = FORTYPE, statespace = STATESPACE,
  randerr = RANDERR, bias = BIAS, p = params, w = matrix_weather,
  calf = init$calendar_fert, calN = init$calendar_Ndep,
  calpT = init$calendar_prunT, caltT = init$calendar_thinT,
  n = NDAYS, NOUT = init$NOUT, sv = STATEVARS, stateers = STATEERS,
  procerr = PROCERR)
}
\arguments{
\item{rs}{Integer, indicating if the model is restarted ( set rs to 1) or started from scratch ( set rs to 0)}

\item{ft}{Integer which indicates the forest type and is set to 1 for coniferous and 2 for deciduous}

\item{bias}{Integer, indicating if the model has a process bias (set bias to 1) or has no process bias ( set bias to 0)}

\item{p}{Vector that contain the model parameters. The order of the parameters in the vector is important and should be consistent with the provided dataset df_params.}

\item{w}{Matrix that contains the model input weather data with dimensions [number of timesteps, number of weather inputs]. The seven weather inputs are: year, day of year, radiation (MJ/day), mean air temperature (deg C), precipitation (mm/day), surface wind speed (m/s) and vapour pressure (kPa)}

\item{calf}{Matrix timeseries of fertilization of dimension [length of timeseries, date vector of length 3]. Where the date vector is made up of (year, day of year, fertilization in kgN /day /m^2)}

\item{calN}{Matrix timeseries of N depostion of dimension [length of timeseries, date vector of length 3]. Where the date vector is made up of (year, day of year, N deposition in kgN /day /m^2)}

\item{calpT}{Matrix timeseries of prunings of dimension [length of timeseries, date vector of length 3]. Where the date vector is made up of (year, day of year, pruning fraction)}

\item{caltT}{Matrix timeseries of forest thinnings of dimension [length of timeseries, date vector of length 3]. Where the date vector is made up of (year, day of year, thinning fraction)}

\item{n}{Length of run in days}

\item{NOUT}{Number of output variables}

\item{sv}{Vector that contains the state variables}

\item{staterr}{Integer, indicating if the model has a statistical error (set staterr to 1) or has no stat. error ( set staterr to 0)}

\item{er}{Vector that contains the error variables}
}
\value{
Returns a matrix of dimension [length of run in days, number of outputs]
}
\description{
Run the forest model BASFOR
}
\details{
This is a wrapper function to run BASFOR.
}
\examples{
FORTYPE        <- as.integer(1)
params         <- df_params[,1]
year_start     <- as.integer(1900); doy_start <- as.integer(1); NDAYS <- as.integer(40543)
df_weather     <- weather_CONIFEROUS_1
matrix_weather <- weather_BASFOR( year_start, doy_start, NDAYS, df_weather )
init           <- initialise()
output         <- run_mod_model()
}
